Final Commit
Commit with updated instructions on installing and running locally. Installing and connecting to the remote host for testing (Not 100% sure what will happen if multiple people connect from their Visual Studio instance to the Azure Hosted App?). Also the web address for the App is included in the README but it will probably be stopped unless a request to start it is made.